// src/app/models/offer.class.ts
export class OfferData {
    id?: number;
    title: string;
    description: string;
    requirements: string;
    salaryAmountMin: number;
    salaryAmountMax: number;
    subscribed: boolean;
    location: string;
    contractTypeId: number;
    workdayId: number;
    requestedProfessionalExperienceId: number;
    yearsOfExperienceId: number;
    businessId: number;
}